+++ include/system/Window.h
@@ -15,8 +15,6 @@
 #include <glm/vec4.hpp>
 #include <glm/gtc/type_precision.hpp>
 
-#include "Shader.h"
-
 class Window {
   public:
     static int initialize();
@@ -35,38 +33,6 @@ class Window {
 
     static void setTextInput(bool t);
     static bool getTextInput();
-
-    static void drawTextGL(std::vector<glm::vec2>& vertices, std::vector<glm::vec2>& uvs,
-                           glm::vec4 color, unsigned int texture);
-
-    static void drawGL(std::vector<glm::vec3>& vertices, std::vector<glm::vec2>& uvs,
-                       std::vector<unsigned short>& indices, glm::mat4 MVP, unsigned int texture);
-
-    static void drawGL(std::vector<glm::vec3>& vertices, std::vector<glm::vec3>& colors,
-                       std::vector<unsigned short>& indices, glm::mat4 MVP, int mode = GL_TRIANGLES);
-
-  private:
-    static int initializeGL();
-    static void shutdownGL();
-
-    static Shader textShader;
-    static const char* textShaderVertex;
-    static const char* textShaderFragment;
-
-    static Shader imguiShader;
-    static const char* imguiShaderVertex;
-    static const char* imguiShaderFragment;
-    friend class UI;
-
-    static Shader textureShader;
-    static const char* textureShaderVertex;
-    static const char* textureShaderFragment;
-
-    static Shader colorShader;
-    static const char* colorShaderVertex;
-    static const char* colorShaderFragment;
-
-    static unsigned int vertexArrayID;
 };
 
 #endif